Various refactors
