--- .vscode/settings.json
@@ -1,8 +1,0 @@
-{
-    "cSpell.words": [
-        "creds",
-        "inferencing",
-        "presigner",
-        "utilise"
-    ]
-}
