+++ package.json
@@ -5,13 +5,12 @@
   "bin": {
     "cdk-example": "bin/reefguide-web-api.js"
   },
-  "main": "src/index.ts",
+  "main": "src/api/index.ts",
   "scripts": {
-    "start": "env-cmd ts-node --files src/index.ts",
+    "start": "env-cmd ts-node --files src/api/index.ts",
     "watch": "tsc -w",
-    "dev": "env-cmd ts-node-dev --files src/index.ts",
+    "dev": "env-cmd ts-node-dev --files src/api/index.ts",
     "build": "tsc",
-    "start:dist": "node dist/src/index.js",
     "lint": "eslint --fix src test",
     "test": "jest",
     "typecheck": "tsc --noEmit",
@@ -39,7 +38,6 @@
     "@codegenie/serverless-express": "^4.15.0",
     "@prisma/client": "^5.22.0",
     "@types/aws-lambda": "^8.10.145",
-    "@types/helmet": "^4.0.0",
     "aws-cdk-lib": "2.158.0",
     "axios": "^1.7.7",
     "bcryptjs": "^2.4.3",
@@ -89,7 +87,6 @@
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.5.3",
-    "webpack": "^5.96.1",
     "webpack-cli": "^5.1.4",
     "webpack-node-externals": "^3.0.0",
     "zod-prisma-types": "^3.1.8"
