+++ src/api/apiSetup.ts
@@ -3,180 +3,24 @@ import express from 'express';
 import helmet from 'helmet';
 import morgan from 'morgan';
 import { passport } from './auth/passportConfig';
-
-import api from '.';
 import * as middlewares from './middlewares';
-
 import { PrismaClient } from '@prisma/client';
 import { config } from './config';
-import { registerUser } from './services/auth';
+import { router as adminRoutes } from './admin/routes';
+import { getJwks } from './auth/jwtUtils';
+import { router as authRoutes } from './auth/routes';
+import { router as jobRoutes } from './jobs/routes';
+import { router as noteRoutes } from './notes/routes';
+import { router as polygonRoutes } from './polygons/routes';
+import { router as userRoutes } from './users/routes';
 
-export const prisma = new PrismaClient();
-
-/**
- * Initializes or verifies admin users in the system.
- *
- * This function ensures that all required administrative users (manager, worker, and admin)
- * exist in the system with the correct roles and credentials. For each required admin user,
- * it will:
- * 1. Check if the user already exists
- * 2. Create the user if they don't exist
- * 3. Verify/update admin roles for existing users
- * 4. Verify successful creation/update
- */
-export const initialiseAdmins = async () => {
-  console.log('=== initialiseAdmins START ===');
-
-  console.log('Config check:', {
-    hasConfig: !!config,
-    hasCreds: !!config?.creds,
-    managerCreds: {
-      hasUsername: !!config?.creds?.managerUsername,
-      hasPassword: !!config?.creds?.managerPassword,
-    },
-    workerCreds: {
-      hasUsername: !!config?.creds?.workerUsername,
-      hasPassword: !!config?.creds?.workerPassword,
-    },
-    adminCreds: {
-      hasUsername: !!config?.creds?.adminUsername,
-      hasPassword: !!config?.creds?.adminPassword,
-    },
-  });
-
-  const initialise: { email: string; password: string }[] = [
-    {
-      email: config.creds.managerUsername,
-      password: config.creds.managerPassword,
-    },
-    {
-      email: config.creds.workerUsername,
-      password: config.creds.workerPassword,
-    },
-    {
-      email: config.creds.adminUsername,
-      password: config.creds.adminPassword,
-    },
-  ];
-
-  console.log('Initialising users count:', initialise.length);
-  console.log(
-    'Users to initialise:',
-    initialise.map(user => ({ email: user.email })),
-  );
-
-  for (const { email, password } of initialise) {
-    console.log(`\n=== Processing user: ${email} ===`);
-
-    try {
-      // First, try to find the user
-      console.log(`Checking if user exists: ${email}`);
-      const existingUser = await prisma.user.findUnique({
-        // Changed from findUniqueOrThrow
-        where: { email },
-        select: {
-          // Only select the fields we need
-          id: true,
-          email: true,
-          roles: true,
-        },
-      });
-
-      if (existingUser) {
-        console.log(`User ${email} already exists:`, {
-          id: existingUser.id,
-          roles: existingUser.roles,
-        });
-
-        // Optionally, verify and update roles if needed
-        if (!existingUser.roles.includes('ADMIN')) {
-          console.log(`Updating user ${email} to include ADMIN role`);
-          await prisma.user.update({
-            where: { email },
-            data: {
-              roles: [...existingUser.roles, 'ADMIN'],
-            },
-          });
-          console.log(`Successfully added ADMIN role to ${email}`);
-        }
-      } else {
-        console.log(`User ${email} not found, creating new user...`);
-        const newUser = await registerUser({
-          email,
-          password,
-          roles: ['ADMIN'],
-        });
-        console.log(`Successfully created new admin user: ${email}`, {
-          success: !!newUser,
-          roles: ['ADMIN'],
-        });
-      }
-    } catch (e: any) {
-      console.error(`\n!!! Error processing user: ${email} !!!`);
-      console.error('Error type:', e.constructor.name);
-      console.error('Error message:', e.message);
-      console.error('Error stack:', e.stack);
-      if (e.code) {
-        console.error('Error code:', e.code);
-      }
-      if (e.meta) {
-        console.error('Error metadata:', e.meta);
-      }
-      // You might want to throw here depending on your requirements
-      throw new Error(`Failed to initialize admin user ${email}: ${e.message}`);
-    }
-  }
-
-  // Final status check
-  try {
-    const adminUsers = await prisma.user.findMany({
-      where: {
-        roles: {
-          has: 'ADMIN',
-        },
-      },
-      select: {
-        email: true,
-        roles: true,
-      },
-    });
-
-    console.log('\n=== Final Admin Users Status ===');
-    console.log('Total admin users:', adminUsers.length);
-    console.log(
-      'Admin users:',
-      adminUsers.map(user => ({
-        email: user.email,
-        roles: user.roles,
-      })),
-    );
-
-    // Verify all required admins are present
-    const allEmails = adminUsers.map(u => u.email);
-    const missingAdmins = initialise
-      .map(u => u.email)
-      .filter(email => !allEmails.includes(email));
-
-    if (missingAdmins.length > 0) {
-      console.error('Warning: Some admin users are missing:', missingAdmins);
-      throw new Error(
-        `Failed to initialize all required admin users. Missing: ${missingAdmins.join(', ')}`,
-      );
-    }
-
-    console.log('✅ All required admin users are properly initialized');
-  } catch (error) {
-    console.error('Failed to verify final admin users status:', error);
-    throw error;
-  }
-
-  console.log('\n=== initialiseAdmins END ===');
-};
 require('dotenv').config();
-
 require('express-async-errors');
+
+// Setup app
 const app = express();
 
+// Helmet and middleware
 app.use(morgan('dev'));
 app.use(helmet());
 app.use(cors());
@@ -184,6 +28,41 @@ app.use(express.json());
 
 // Passport authentication
 app.use(passport.initialize());
+
+export const prisma = new PrismaClient();
+
+// Setup the /api sub router
+const api = express.Router();
+
+// jwks.json wkt endpoint
+api.get('/.well-known/jwks.json', (req, res) => {
+  res.json(getJwks());
+});
+
+/** Health check GET route */
+api.get('/', (req, res) => {
+  res
+    .json({
+      message: 'API healthy.',
+    })
+    .send();
+});
+
+// Passport auth routes
+api.use('/auth', authRoutes);
+api.use('/polygons', polygonRoutes);
+api.use('/notes', noteRoutes);
+api.use('/admin', adminRoutes);
+api.use('/users', userRoutes);
+api.use('/jobs', jobRoutes);
+
+const port = config.port || 5000;
+
+app.listen(port, () => {
+  /* eslint-disable no-console */
+  console.log(`Listening: http://localhost:${port}`);
+  /* eslint-enable no-console */
+});
 
 // API base router
 app.use('/api', api);
