+++ src/api/types/errors.ts
