+++ src/infra/components/jobs.ts
@@ -34,9 +34,6 @@ export interface JobSystemProps {
   // Domain and auth
   apiEndpoint: string;
 
-  // TODO use different form of creds...
-  apiAuthToken: string;
-
   // Capacity manager configuration
   capacityManager: {
     cpu: number;
@@ -133,6 +130,8 @@ export class JobSystem extends Construct {
         healthCheck: {
           command: [
             'CMD-SHELL',
+            // This magic redirects the output of health checks into logs for
+            // the container so if it fails we can actually see why (facepalm)
             `curl -f http://localhost:${config.serverPort}/health >> /proc/1/fd/1 2>&1 || exit 1`,
           ],
           interval: Duration.seconds(30),
@@ -210,10 +209,6 @@ export class JobSystem extends Construct {
       command: ['npm', 'run', 'start-manager'],
       image: ecs.ContainerImage.fromAsset('.', {
         buildArgs: { PORT: '3000' },
-        // Very annoying if the docker build includes the infra folder because
-        // then literally any change will cause a new docker image to be
-        // deployed
-        exclude: ['src/infra'],
       }),
       logging: ecs.LogDrivers.awsLogs({
         streamPrefix: 'capacity-manager',
