+++ src/infra/components/reefGuideAPI.ts
@@ -182,9 +182,7 @@ export class ReefGuideAPI extends Construct {
     this.fargateService = new ecs.FargateService(this, 'reefguide-service', {
       cluster: cluster,
       taskDefinition: reefGuideTaskDfn,
-      // TODO Update if we need to
-      // TODO restore
-      desiredCount: 0,
+      desiredCount: 1,
       securityGroups: [serviceSecurityGroup],
       assignPublicIp: true, // TODO Change this if using private subnets with NAT
       // give plenty of time
